refactor: move tests to subdirectories
