+++ src/api/routes/__tests__/inscriptions.test.ts
@@ -1,7 +1,7 @@
 import { cycleMigrations } from '@hirosystems/api-toolkit';
-import { buildApiServer } from '../src/api/init';
-import { MIGRATIONS_DIR, PgStore } from '../src/pg/pg-store';
-import { TestChainhookPayloadBuilder, TestFastifyServer } from './helpers';
+import { buildApiServer } from '../../init';
+import { MIGRATIONS_DIR, PgStore } from '../../../pg/pg-store';
+import { TestChainhookPayloadBuilder, TestFastifyServer } from '../../../../tests/helpers';
 import {
   BitcoinInscriptionRevealed,
   BitcoinInscriptionTransferred,
